Ensure sendExtraNotifyImm scheme is triggered only in Spray mode by adding condition
Summary:
Building on the context from D83756684: Previously, sendExtraNotifyImm was enabled by default since we only supported Spray mode. This diff adds explicit conditions so that sendExtraNotifyImm is only triggered in Spray mode. With this change, sendExtraNotifyImm will not be activated when DQPLB mode is introduced in the future.

Files changed in this diff:
1. ibverbx/Ibverbx.h: Added the sendExtraNotifyImm field to VirtualWc and VirtualCqRequest to indicate if an extra Notify IMM message should be sent. This is now set based on the send work request opcode (IBV_WR_RDMA_WRITE_WITH_IMM) and the load balancing scheme (Spray).

2. ibverbx/tests/IbverbxTest.h: Updated tests with the new sendExtraNotifyImm field.

Reviewed By: cenzhaometa, YulunW

Differential Revision: D83758753

fbshipit-source-id: c25d95bb05533f566d690bf984159474d91f80b0