Remove the unnecessary dynamic link interface
Summary:
For some historical reason, we added those dynamic loader interface for shared library support, before we had the dynamic registration that Tristan implemented. With the dynamic registration, we can keep the static library in BUCK and still have TorchComms work with various backends.
This shared library approach had some drawbacks in the case of HCCL, which is why we turned away from it. And for other backends, this is not being regular tested and removing that support for now makes sense. We can add it back if needed for specific cases in the future.

Reviewed By: rsdcastro

Differential Revision: D85348860

fbshipit-source-id: 3ae14ce7b1a48ec7c81ee81ec73a465822fead3f