Set the name of thread for ease of debugging
Summary:
We spin up many threads, and name them. The names appear in the logs which makes
it easier to correlate. However if we inspect things in GDB or `ps -H -p <>` we
do not see correlation of thread-id to their names.

Explicitly set the name of thread so that we can correlate stack traces or CPU
utilization in various debugging tools.

Reviewed By: arttianezhu

Differential Revision: D85452338

fbshipit-source-id: 5071b46eb72c75708d092eca14cf24aa74158bd6