Omit CudaGraph collectives
Summary: Previously, we will terminate the program if CollTrace encountered cuda graph. This is to ensure we do not accidentally miss any collective in Cuda Graph by CollTrace. However, we later found that support for Cuda Graph is non-trivial and we still need more time to make it happen. Therefore, we simply omit them for now to enable using new colltrace.

Reviewed By: tanquer

Differential Revision: D84476752

fbshipit-source-id: d89ed496a35da9e5f3de418cd29052005342e043