Rebasing nccl 2.28 with latest changes.
Summary:
These diff stack, implements the required updates to rebase NCCL 2.28 for NCCLX with CTRAN integration. It incorporates all changes introduced in version 2.27, applying them on top of NCCL’s latest stable release (2.28). The primary objective of this diff is to enable CTRAN support for NCCLX under NCCL 2.28, ensuring compatibility and leveraging the latest enhancements from the upstream release. In addition, it includes necessary checks inside FastInitTest so that it can ensure the ctran enabling as expected.

Specifically in this diff, all rebasing tasks are implemented.

Reviewed By: zhiyongww

Differential Revision: D85349488

fbshipit-source-id: 370f26f555a224db86c35e03dcf4f4133460c52b