Updating commSpec so that it can support v2.28.
Summary:
These diff stack, implements the required updates to rebase NCCL 2.28 for NCCLX with CTRAN integration. It incorporates all changes introduced in version 2.27, applying them on top of NCCL’s latest stable release (2.28). The primary objective of this diff is to enable CTRAN support for NCCLX under NCCL 2.28, ensuring compatibility and leveraging the latest enhancements from the upstream release. In addition, it includes necessary checks inside FastInitTest so that it can ensure the ctran enabling as expected.

In this specific diff, commspecs and related datatypeconv logics are updated so that it can support both v2_27 and v2_28 of nccl.

Reviewed By: dboyda, zhiyongww

Differential Revision: D85349491

fbshipit-source-id: c16d924d7e75ae5060b16b97842d2535fd3bd790