Make FastInitTest test Ctran Init
Summary:
These diff stack, implements the required updates to rebase NCCL 2.28 for NCCLX with CTRAN integration. It incorporates all changes introduced in version 2.27, applying them on top of NCCL’s latest stable release (2.28). The primary objective of this diff is to enable CTRAN support for NCCLX under NCCL 2.28, ensuring compatibility and leveraging the latest enhancements from the upstream release. In addition, it includes necessary checks inside FastInitTest so that it can ensure the ctran enabling as expected.

In this specific diff, I include fast init test with the required codes so that it can test Ctran init.

Reviewed By: dboyda

Differential Revision: D85065685

fbshipit-source-id: aee990de51be0ff1f9eb28640da1f19f323b4498