Remove completed_queue for NCCL, MCCL & HCCL
Summary:
Similar to D85455174, remove completed_queue for other CCLs.

Context: The original purpose was to prevent destruction outside the main thread of the work object. However now that we have a finalize method, we can rely on the contract that the user will hold onto the object until finalize is called.

Reviewed By: siyengar

Differential Revision: D85771546

fbshipit-source-id: f28e89b8c1fe769f6a144f9684ac4b39f4a44460