Add stack trace to the last error message
Summary: Currently, NCCL Error is not well-surfaced to the MAST UI, making it confusing for the user to understand what is the root cause of an error. Previously we fixed the lastError surfacing in D85412934. This diff further improves the error surfacing by adding NCCL "stack trace" information into it.

Reviewed By: minsii

Differential Revision: D85920836

fbshipit-source-id: 46204e0a39c8de9f86326fe2bac815afea8197ba