change holder of work to intrusive ptr from shared ptr
Summary:
Change holder type of Work objects to an instrusive_ptr.
This will reduce the overhead of pybind return value wrapping logic.
Previously it need to hold a shared_ptr and wrap it. intrsuive_ptr does not
need this wrapping any more. This will bring this overhead closer to the processgroup.

Differential Revision: D85934777