change holder of work to intrusive ptr from shared ptr (#37)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/torchcomms/pull/37

Change holder type of Work objects to an instrusive_ptr.
This will reduce the overhead of pybind return value wrapping logic.
Previously it need to hold a shared_ptr and wrap it. intrsuive_ptr does not
need this wrapping any more. This will bring this overhead closer to the processgroup.

Reviewed By: d4l3k

Differential Revision: D85934777

fbshipit-source-id: 12ab355a76da6ac09fce0553a37a35a734c96a03