Fix cuda graph UTs
Summary:
When doing cuda graph tests, we always switch to a non-default CUDA stream to capture operation and launch it to that stream. However, we set up our input & expected output on the default stream in an async way (`at::Tensor::fill_` is an async operation). Therefore, it is possible that the non-default CUDA stream execute the collective communication before the input/output setup, which cause the test failure.

This diff fix that gap by moving input/output setup after the non-default CUDA stream switch, then all operations will be executed sequentially on the same stream.

Reviewed By: pavanbalaji

Differential Revision: D86485186

fbshipit-source-id: deb5910b994817dceddbf70373e268a1a3339f8b