Fix trace logger test
Summary: Followup to D86586879. UT did not previously end trace record which meant that end time is not set. This means duration underflows and becomes a very large number that parseJson in UT catches and fails. Fix is to set end time correctly and compare values to ints instead of strings since this was changed in D86586879.

Reviewed By: minsii

Differential Revision: D86678693

fbshipit-source-id: 23745934f78c0edbbe7352873eb9b3df1eeb7077