fix NCCLX cvars ASAN abort in folly::split
Summary: Not sure why exactly there is an ASAN abort. But it blames to {D85101083} which moved the calls to the vector intrinsics into the header. Meaning that calls to `folly::split` used to be compiled for the cpu and now, if they are called from a source file compiled for the gpu, these calls to `folly::split` will also be compiled for the gpu. And that might not be compatible with AVX intrinsics.

Differential Revision: D86726139

fbshipit-source-id: ec289746d9d13a4331d936ec62208319e4cb63d0