Integrate RCCLX in setup.py and tweaks to Cmakelists
Summary:
This adds integration to setup.py for RCCLX backend, and some tweaks to CMakeLists to take USE_RCCLX parameter to build the backend and install it.

Also add the run_integration_tests script for rcclx backend and a setup_rcclx.sh. Once D84125652 is in, I will have a separate build_test.yaml for RCCLX on mi350_runner.

Reviewed By: siyengar

Differential Revision: D86221377

fbshipit-source-id: 04a9f8d14debbad67cf05041433306956b8661ce