create bootstrap::AbortableSocket skeleton from bootstrap::Socket
Summary:
# Summary
-------------

This diff introduces the `AbortableSocket` and `AbortableServerSocket` classes. These classes are **based heavily** off the existing `comm/ctran/bootstrap/Socket.(cc|h)` (i.e., `AbortableSocket` reuses a lot of the code from the existing `Socket` class.)

`ctran::utils::Abort` is integrated directly with `AbortableSocket` and `AbortableServerSocket` to support aborting operations gracefully.

For now, we rely on destructors to handle shutting down sockets. The exception to this is in `AbortableSocket::connect` -- if there's an error while attempting to connect, then the socket will be `close()`.

This first diff establishes the general skeleton with the `ctran::utils::Abort` dependency and basic resource management for the sockets. Specifically, it includes:
*   getAbort()
*   CHECK_ABORT_RETURN()

# (Sub)Stack Overview
*   D86240195 introduces updated socket connection logic.
*   D86240199 introduces updated data transfer operation implementations.
*   D86240196 completes the `AsyncServerSocket` implementation.

Reviewed By: arttianezhu

Differential Revision: D86233995

fbshipit-source-id: 97d00489ab54664baf1d0c8940a3256c03d3b594