AbortableSocket: Socket Connection Impl
Summary:
This diff implements the core logic for connection establishment by updating the old `Socket.cc` implementation for `AbortableSocket.cc`.

In particular, this diff updates the implementation for the following methods:
*   `connect()` method
*   `prepareSocket()` private method
*   `waitForConnect()` private method
*   `waitForEvent()` private method
*   `waitForWritable()` method (needed by `waitForConnect`)

Connection is the first major operational capability for `AbortableSocket` and is a natural checkpoint before introducing data transfer operations. All the waiting/polling infrastructure is introduced in this diff.

Reviewed By: arttianezhu

Differential Revision: D86240195

fbshipit-source-id: d9f5144168c4a5eaa37f66f8d9f47e4ea9a0fd35