ctran::utils::Abort add Timeout state and related APIs
Summary:
This diff introduces a `TimedOut()` method to check if the timeout set on an `Abort` as expired, and a `HasTimeout` method to check if the `Abort` has had a timeout set on it. This also updates the `Test` method to explicitly set the `abort_` and `timedOut_` flags if it finds that the timeout has expired. Finally, this adds a `TimeRemaining` method, which returns an `std::chrono::milliseconds` indicating the time remaining until the Abort's timeout expires.

These updates are performed to prepare the `ctran::utils::Abort` class for use/integration with a new `ctran::bootstrap::AbortableSocket` to enable `CtranIb` bootstrapping fault tolerance / cancelability.

Reviewed By: arttianezhu

Differential Revision: D85820276

fbshipit-source-id: db4175dcd71dc761641326b622d8ae301dee8b1b