Change Coordinator instance to folly::Singleton
Summary: Previously, Coordinator was a member of IbvDevice and passed as a pointer to IbvPd, IbvVirtualQp, and IbvVirtualCq. To decouple IbvVirtualQp and IbvVirtualCq from IbvDevice, this diff removes the direct Coordinator dependency. The Coordinator is now a folly::Singleton, so the same instance is shared by all IbvVirtualQp and IbvVirtualCq, even across different devices.

Reviewed By: cenzhaometa

Differential Revision: D85995440

fbshipit-source-id: 25f56e249072c4e01ca261c380b9d554cb7fee34