Introduce signal/checkSignal in MultiTbSync util
Summary:
This patch adds a signal-checkSignal pair in multi-thread-block sync module. This can be typically used for producer-consumer scheme.

For instance
```
if (producer) {
   // update shared data
   signal(cnt, 1);
}

if (consumer) {
   while(checkSignal(cnt, 1));
   // read shared data
}
```

In real algorithm, the consumer side is usually nonblocking, i.e., consumer can do something else if the current signal is not yet posted.

Optional configs:
- The producer side signal() can be with internal __syncthreads(), or without if callsite already ensures cross-threads sync within the producer block.
- A signal can be checked by multiple consumers in a multi-cast style. See extended UT.

Reviewed By: dsjohns2

Differential Revision: D87007403

fbshipit-source-id: ebbf9966da3cedf116d9d774c95481ddb5a3683c