Add memory registration before hook registration
Summary: When it comes to mem pool registration. After the singleton is created and callback is registered, all memory will be tracked. But we also want to take care of the memory allocation happens before this registration moment, that is why we add this registration and only call it once when we create the singleton.

Reviewed By: siyengar

Differential Revision: D86907322

fbshipit-source-id: d51f6616f4d4e97696f04c8bf760b3b7302d2d00