Removing AllGather algo check from test setup
Summary: Previously, in the test setup in the CtranDistAllgatherTests file, the test setup involves setting the algo for AllGather and checking if it is compatible by calling the ctranAllGatherSupport function. This ends up being unnecessary since in the tests themselves, the algo is overwritten. In this diff, this part of the test setup is removed to avoid the repetition.

Reviewed By: Regina8023

Differential Revision: D87367058

fbshipit-source-id: 3b9b639915c6e6a487c576e485dc20b82a8006ed