Adding algo to CtranAllGather API
Summary: In a previous diff (D85789334), the ctranAllGatherSupport function was amended to check if a particular algorithm works with AllGather and the number of local ranks for the given environment. In this diff, we change the CtranAllGather API to take algo as an argument since we now do an algo check prior to calling the API and do not want to override with a different algorithm. Also this avoids repetitive code by only checking if the algo is compatible one time each time we call AllGather.

Reviewed By: Regina8023

Differential Revision: D87284532

fbshipit-source-id: 73490856498e1f97aa1d081b2db2331805e5f880