Integrating AsyncSocket with CtranIb
Summary:
This diff integrates the new `AsyncSocket` and `AsyncServerSocket` classes with `CtranIb`. In doing so, this diff helps to enable/unblock our ongoing FT efforts by unblocking subsequent work to make MCCL's `init` to be FT.

This diff exposes the `bootstrapAccept` and `bootstrapConnect` methods as part of the public API of `CtranIb`. This is done primarily to enable writing unit tests involving these methods, which has been done extensively in D86250310, D86251415, D86251416, D85962818, D85974129, D85974483, and D85963054. Thorough testing of the `CtranIb` class and the various concurrent and failure/abort scenarios that can be encountered necessitates the capability to directly call `bootstrapAccept` and `bootstrapConnect`.

It is also worth noting that there are other relatively similar APIs within `CtranIb` that are already public, such as `connectVc` and `connectVcDirect`.

# Specifying `ISocketFactory` Implementation to `CtranIb`

The `CtranIb` constructors accept an `ISocketFactory` argument; however, this will typically not be accessible to public APIs. Instead, `CtranIb` uses the enabled status of the `ctran::utils::Abort` argument (passed to constructors). If `abortCtrl_->Enabled()`, then `CtranIb` will create and use an `AbortableSocketFactory`; otherwise, `CtranIb` will fall back to old `Socket` implementation and use `SocketFactory`.

# Behavior Change in `CtranIb::bootstrapAccept`

When FT (i.e., the `CtranIb::abortCtrl_`) is enabled, `bootstrapAccept` will no longer throw an `std::runtime_error` whenever an error is returned by a socket operation. Instead, errors are only thrown for non-FT-related errors (i.e., we don't want to throw an uncaught `std::runtime_error` simply because an Abort was triggered locally or on a remote peer).

Reviewed By: arttianezhu

Differential Revision: D85155366

fbshipit-source-id: 55b3f3450ffa2d65babc504979b776d9c8916840