MockISocket, MockIServerSocket, and CallbackInjectorSocketFactory for fault injection and test orchestration
Summary:
Introduces a `MockISocket` and `CallbackInjectorSocketFactory`, the latter of which is a simple and concise `ISocketFactory` implementation that links `AbortableSocket` instances with a `MockISocket` that allows for failure injection at function boundaries.

The `CallbackInjectorSocketFactory` also fits well into the existing dependency injection infrastructure of `CtranIb` and enables complete testing of production codepaths.

This is required because `CtranIb` dynamically creates `ISocket` instances internally, so they cannot simply be mocked and passed into `CtranIb`. (If so, then we could use `EXPECT_CALL` with mocked `ISocket` instances to inject callbacks for failure injection and whatnot.)

Reviewed By: arttianezhu

Differential Revision: D85711746

fbshipit-source-id: f97c0e0070559a48184b00d377697fa1ff69cbdd