torchcomms: enable typing
Summary:
This enables a lot more of Pyre typing on torchcomms targets. This is a prereq to OSS typing CI.

This has a few type fixes but I've limited it to just `-> None` and removing old incorrect type warnings.

A few files have been switched to pyre-strict since they pass.

Reviewed By: fduwjj, tanquer

Differential Revision: D87815835

fbshipit-source-id: 123417cc38b5e0d25bbef3abe2ff8f95c77f0826