Refactor combine API: change recvbuffs to recvbuff
Summary:
For API combine, the recvbuff will only be a single pointer pointing to the start of the array.

Note: to avoid over-eng and for implementation simplicity, we choose to support the two APIs by keeping both recvbuffs (for dispatch) and recvbuff (for combine) in ctranAlltoallvDynamicSplitNonContig, as we now only have two moving variables for the two APIs: recvbuffs and recvAllSplitLengths.
In the future, if the moving variables increase and it requires more extendibility, we should implement a new class/struct (and use union/variant for recvbuff/recvbuffs) like Hints to support various metadata type.

Reviewed By: cenzhaometa

Differential Revision: D87719456

fbshipit-source-id: 0682de59585654864e2e50b3ed338e6a6f72ff07