Updating ctranReduceScatterRing algo to handle recvcount=0
Summary: Previously, running ctring for ReduceScatter with 2x1 topology (which is the algorithm run for ctranReduceScatter when the topo is 2x1) produced a divide by 0 error. In this diff, we add a check for if recvcount is 0, in which case we set the number of rounds to be 1 in the ring configuration instead of doing a calculation resulting in division by 0.

Reviewed By: Regina8023

Differential Revision: D87692951

fbshipit-source-id: 27d21d49702450749a75ccb346460d29263aa0f3