Create prepare benchmark
Summary:
This patch defines a standalone benchmark for dedup internal prepare kernel. This allows us to catch any bottleneck in prepare and finetune.

Two changes to the prepare algorithm:
- To profile overhead of each prepare subroutine, we refactored prepare kernel to optionally control subset of roles to run.

- To run the benchmark with single GPU env, we define `CommStateXMock` and `ExecArgsMock` and allow to be passed via `launchExecResetForTest()`

- Benchmark result suggest numThreads 256 is generally better than 128. And 512 may outperform 256 only at 8 nodes. This patch defines CVAR to control prepare num threads and set to 256 by default.

- For further optimization, consider assign more thread blocks for the bottleneck preparations (see test plan results)

TODO: This benchmark should be converted to folly benchmark. But it is still unclear to me how to assign multiple parameters together with user counter. Besides, since we want to scan combination of parameters, we need a way to do so with folly::benchmark, similar to `testing::Combine()` from gtest.

Reviewed By: dsjohns2

Differential Revision: D87659315

fbshipit-source-id: 594f09e037c31b79e8d8b42f3098ad521a7b9dd3