Refactor param and EnvRAII assignment
Summary:
Dedup UT contains too many parameters and hard to read. In addition, the worker group parameters need to be assigned to CVAR via EnvRAII at each test, which is redundant.

This patch made two changes to improve the readability:
1. Define meaningful param structs
2. Define common macro to assign EnvRAII

Besides, this patch reduces number of blocks to 4000 in some test cases to reduce run time. This should be sufficient to cover multi-step cases

Reviewed By: dsjohns2

Differential Revision: D88127049

fbshipit-source-id: a2bdc29981eb3e810c8e474eecd40a2934c9f4f1