Temporary work-around for treating prefixed_stringlist CVARs as string
Summary:
# Context
-----------
There's an issue with one of the CVARs for NCCLX v2.28:

`NCCL_IB_HCA` is defined as a `prefixed_stringlist` in the `nccl_cvars.yaml` file, which leads to the creation of `NCCL_IB_HCA` and `NCCL_IB_HCA_PREFIX` C++ variables. These are used in various places, including in CTRAN and CTRAN UTs.

However, in NCCLX v2.28, we have `ncclGetEnv("NCCL_IB_HCA")`, expecting it to be a string. We then try to manually parse that value in NCCLX. If we change it to be defined as a string, then it'll break a bunch of places that expect `NCCL_IB_HCA_PREFIX` to exist.

# Workaround
----------------
As a temporary workaround, we now also create a `XYZ_STRINGVALUE` C++ variable for `prefixed_stringlist` CVARs, and have `nccl_baseline_adapter::ncclGetEnvImpl` fall back to checking the string map for `XYZ_STRINGVALUE` (i.e., treating it like a string).

We can use this to unblock NCCLX v2.28 while we decide on a long-term solution for handling `prefixed_stringlist` CVARs.

Reviewed By: zhiyongww

Differential Revision: D88186708

fbshipit-source-id: 15086c772840fa594a36424ef3f633f583f5f7ae