error handling on mcclComm::init
Summary:
Arguably, the capture scope is very wide. The reason is that some of the errors should be crashing quickly, while others should be retried.

I found a really small number of cases, where `std::runtime_error` is used in place of something more suitable, such as an `invalid_argument`. We will follow up to clean these up, so that non-recoverable items would fail fast and recoverable ones are handled by Mccl. But these findings are harmless for initial enablement.

Reviewed By: dboyda

Differential Revision: D85824307

fbshipit-source-id: 1c5b4e647ccb3ee7f2eaad927071dd8f40411b89