torchcomms/dummy: enable using dummy backend outside of tests (#75)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/torchcomms/pull/75

This enables using the dummy backend outside of tests. This allows for mocking out the distributed parts of a model for local performance testing.

This also does some small cleanups/registration to make the Dummy backend more similar to other backends.

Meta:

This required significant changes to how we package the backend .so files to avoid duplicate library linking when building with BUCK.

https://docs.google.com/document/d/1A1_djQlNcTznXm72G7R_MELW-j0S05pz4WTFSpeDdts/edit?tab=t.0#heading=h.7ws522vzwljf

Reviewed By: fduwjj

Differential Revision:
D88218584

Privacy Context Container: L1397144

fbshipit-source-id: 464d207c38e589f0c9cf39a17f101a15af09f162