Updating ctranReduceScatter to take algo as parameter
Summary: Previously, the ctranReduceScatter API did not take algo as a parameter. In this diff, the function is updated to accept algo as an argument in both the API function and the function that checks whether ctran is supported. The support function will now not only check if ctran in general is supported, but whether the specific algorithm is supported. This also allows future overwriting of ReduceScatter algo if necessary.

Reviewed By: Regina8023

Differential Revision: D88208509

fbshipit-source-id: e55fc65fce74e6f35730d157929dc430f387f65d