Updating ctranBroadcast to take algo as parameter
Summary:
Part of larger effort to add algo as a parameter for all collective APIs, allowing for future overwriting of algo as necessary and also allowing for consistency across collectives.

Prior to this diff, ctran Broadcast did not take algo as a parameter in its API or in its support check function. This diff adds algo as a parameter to those functions and updates all the places that call those functions accordingly.

Since all algo options work for all topologies in the case of broadcast, the support check function does not have any code that specifically checks for algo support. In the future if a new algo is added as an option for broadcast, the support check function should be updated as well.

Reviewed By: Regina8023

Differential Revision: D88318370

fbshipit-source-id: 85c0167b77e44a1d55b1ebddc047ae362b1e7122