Improve error message for unsupported NCCL mem allocator
Summary:
Updated error messages in TorchCommNCCL and TorchCommNCCLX to include
the CUDART version number when the NCCL memory allocator is not
supported. This provides better diagnostic information to users,
making it clearer which CUDA runtime version is causing the
incompatibility. The error message now displays "NCCL mem allocator is
not supported in CUDART version %d" instead of the generic "not
supported in this NCCL version" message.

Reviewed By: siyengar

Differential Revision: D88564927

fbshipit-source-id: 2daf1f8bc6435ba0612af8c25ee5e17ad7a5f65f