Add deviceCnt_ member in IbvVirtualQp
Summary: To support the optional deviceToKeys mapping field in IbvVirtualQp::postSend (in a future diff), users must provide this field when IbvVirtualQp spans multiple NIC devices (i.e., the vector of IbvQp includes QPs from different NICs, such as on the GB200 platform). To enable proper error handling when the field is omitted in these cases, we introduce a deviceCnt_ member in IbvVirtualQp to track the number of NIC devices involved.

Reviewed By: cenzhaometa

Differential Revision: D88286256

fbshipit-source-id: 59c93a740476cb0e2871ac747e751f32e94ba63a