Add deviceId_ as unique IbvDevice identifier and corresponding logic to assign it
Summary:
In the GB200 platform, multiple NIC devices may share the same physical QP number. To ensure accurate mapping between physical QP numbers and their corresponding virtual QPs, this diff introduces an IbvDevice identifier within the IbvDevice class. This enhancement allows the combination of the physical QP number and the IbvDevice identifier to serve as a unique key for identifying a specific virtual QP.
Following D86119069, this diff adds the deviceId_ field as a unique identifier for each IbvDevice and implements the necessary logic to assign this identifier. Additionally, multi-threaded unit tests are included to validate the new functionality.

Reviewed By: cenzhaometa

Differential Revision: D88219380

fbshipit-source-id: 574658f41b3d2bb35c827c8ce8cb19a412e34ffa