CtranIb: replace std::runtime_error with ctran::utils::Exception
Summary:
As described in T238821628, we need to perform an audit of how errors are handled and reported within CTRAN.

To this end, `ctran::utils::Exception` is preferred to `std::runtime_error`. This diff replaces the explicit instances within `CtranIb` in which an `std::runtime_error` is thrown with a `ctran::utils::Exception`.

Reviewed By: arttianezhu

Differential Revision: D87866984

fbshipit-source-id: 5467da8584ca6500e8fb3f16fe44881dba025781