Replace uses of FB_COMMCHECKTHROW with FB_COMMCHECKTHROW_EX in CtranIb
Summary:
As described in T238821628, we need to perform an audit of how errors are handled and reported within CTRAN.

To this end, `ctran::utils::Exception` is preferred to `std::runtime_error`.

This diff replaces usages of `FB_COMMCHECKTHROW` (which throws an `std::runtime_error`) with `FB_COMMCHECKTHROW_EX` (which throws a `ctran::utils::Exception`) within `CtranIb.cc` and `CtranIb.h`.

In order to achieve this, the `CtranIb` public API required two new additions: `getRank()` and `getCommHash()`. These enable the `rank` and `commHash` values to be obtained by `CtranIbEpochRAII`, which uses the `FB_COMMCHECKTHROW_EX ` macro.

Reviewed By: arttianezhu

Differential Revision: D87869921

fbshipit-source-id: 03f1a2c4e60ffc89a5df1db93246d0f9ea4c97ed