Updating ctranAllToAll to take algo as parameter
Summary:
Part of larger effort to add algo as a parameter for all collective APIs, allowing for future overwriting of algo as necessary and also allowing for consistency across collectives.

Prior to this diff, ctran AllToAll did not take algo as a parameter in its API or in its support check function. This diff adds algo as a parameter to those functions and updates all the places that call those functions accordingly.

Since there is only one algo option for ctranAllToAll (algo=ctran) and it works for all topologies, the support check function does not have any code that specifically checks for algo support. In the future if a new algo is added as an option for AllToAll, the support check function should be updated as well.

Reviewed By: Regina8023

Differential Revision: D88531654

fbshipit-source-id: 11fd68643fd139616c632133a9ae6c14bef2ebd6