Consolidate logGpuMemoryStats Into CtranTestUtils
Summary:
The rule of thumb is:
- If the `util/helper` is ctran-specific (i.e., only used under `ctran/`), then we should put it in `CtranTestUtils.h` and `CtranTestUtils.cc`.
- Otherwise, we should put it in `TestUtils.h`.

The `logGpuMemoryStats` function was defined in two places, namely `CtranTestUtils.h` and `testinfra/TestUtils.h`. All usages of `logGpuMemoryStats` were within the `comms/ctran` directory, so this diff consolidates the function accordingly.

This diff also updated the definition of `logGpuMemoryStats` in `CtranTestUtils` to include MB output (matching the richer version that was in `TestUtils.h`). The new implementation logs both bytes and MB values for better readability.

I removed the inline `logGpuMemoryStats` function since it's now consolidated in
`CtranTestUtils.cc`, and all usages were changed to `ctran::logGpuMemoryStats`.

Reviewed By: cenzhaometa

Differential Revision: D88594722

fbshipit-source-id: eacc0d3cbdc5116d2c607b659aff340b486ba28b