Move TcpStorePhase from CtranXPlatUtUtils to TestDistUtils
Summary:
This diff moves `TcpStorePhase` from `CtranXPlatUtUtils .h` to `TestDistUtils.h`, as `TestDistUtils.h` is the only place that uses `TcpStorePhase` and the related bits of code.

Ultimately, `CtranXPlatUtUtils.h` will be removed entirely in subsequent diffs, and we'll likely rearrange some of the stuff in `TestXPlatUtils.h`, as some may make more sense to be within the CTRAN library itself.

Reviewed By: cenzhaometa

Differential Revision: D88513996

fbshipit-source-id: 0d72379eff6db3b315b075b0952de60d8cfb1d47