Port CtranXPlatUtUtils::logGpuMemoryStats to CtranTestUtils
Summary:
This diff moves the `logGpuMemoryStats` function from `CtranXPlatUtUtils.(cc|h)` to `CtranTestUtils.(cc|h)`.

## Future Work
There is another function of the same name (and with the same declaration) defined in `comms/testinfra/TestUtils.h` and used in several other files including:
- `CtranDistCommUT.cc`
- `memCacheAllocatorTests.cc`
- `CtranMapperRegMemBench.cc`
- `SlabAllocatorTest.cc`
- `memoryUtilsTests.cc`

So, there is a question of whether we'd like to consolidate the two functions or not.

Reviewed By: cenzhaometa

Differential Revision: D88500981

fbshipit-source-id: 53065abc2e66b43580dc657c13d8deb41ed711f6